pyproject.toml
[project]
name = "animal-medical-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.3.3",
    "python-decouple>=3.8",
]
